test: Configure coverage report
Configure Jest's coverage report so that tests fail if coverage drops
bellow the configured thresholds.

In order to not block a developer from pushing code in development to
a remote server, the coverage requirements are not enforced in `yarn
test`, and `yarn test:coverage` must be used to have the report.