chore(shell): Add internal API for navigation
Add an internal API to serve the navigation items to be shown on the
application menu. This will allow us to have a UI partially served by
the backend.

We cannot include the icons in the navigation response as bare dynamic
imports are not working correctly with react-icons, even when using Vite
plugins for that effect.

Issue #400